requirements.txt
# (GPU) RTX5090
--extra-index-url https://download.pytorch.org/whl/nightly/cu128
torch --pre

numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
tqdm>=4.62.0

torchvision
torchaudio
albumentations

# GPU-based augmentations and visualization
kornia>=0.7.0
matplotlib>=3.5.0

ultralytics>=8.3.0
